Split IncomingMessageProcessor from CommsProcessor
