+++ core/commons/src/main/java/com/bezirk/middleware/core/comms/Comms.java
@@ -26,7 +26,6 @@ import com.bezirk.middleware.core.comms.processor.EventMsgReceiver;
 import com.bezirk.middleware.core.control.messages.ControlLedger;
 import com.bezirk.middleware.core.control.messages.ControlMessage;
 import com.bezirk.middleware.core.control.messages.EventLedger;
-import com.bezirk.middleware.core.control.messages.Ledger;
 
 /**
  * This class is created to de-couple from the rest of bezirk (spheres / sadl / stream)
@@ -37,44 +36,6 @@ import com.bezirk.middleware.core.control.messages.Ledger;
  * </p>
  */
 public interface Comms {
-
-//    /**
-//     * start the communication
-//     */
-//    boolean startComms();
-//
-//    /**
-//     * stop the communication
-//     */
-//    boolean stopComms();
-//
-//    /**
-//     * close the communication
-//     */
-//    boolean closeComms();
-//
-//    /**
-//     * restart the underlying comms
-//     */
-//    boolean restartComms();
-
-    /**
-     * Set the sphere for sadl. for late initialization
-     *//*
-    void setSphereSecurity(final SphereSecurity sphereSecurity);*/
-
-    /**
-     * TODO: Split the interface for controlling comms component as CommsCtrl
-     * and below access related as Comms
-     * */
-
-    /**
-     * send the control or event message depends of ledger type
-     */
-    @Deprecated
-    // Use sendEventLedger, or sendControlLedger
-    boolean sendMessage(Ledger message);
-
     /**
      * Send event ledger
      */
@@ -88,23 +49,9 @@ public interface Comms {
     boolean sendControlMessage(ControlMessage message);
 
     /**
-     * @param notification
-     * @return
-     */
-    boolean registerNotification(CommsNotification notification);
-
-    /**
      * this is on each comms instance returns its own created id
      */
     String getNodeId();
-
-    /**
-     * Initialize the communications
-     * creates queues, threads, sockets
-     **/
-//    boolean initComms(CommsProperties commsProperties, InetAddress addr,
-//                      SphereSecurity sphereSecurity,
-//                      Streaming streaming);
 
     boolean registerControlMessageReceiver(ControlMessage.Discriminator id, CtrlMsgReceiver receiver);
