+++ core/commons/src/main/java/com/bezirk/middleware/core/comms/CommsNotification.java
@@ -25,7 +25,6 @@ package com.bezirk.middleware.core.comms;
 import com.bezirk.middleware.core.control.messages.MessageLedger;
 
 /**
- * Created by vnd2kor on 12/9/2015.
  * Notification from comms to platform specific bezirk middleware application regarding
  * diagnostic, error messages
  */
@@ -39,9 +38,4 @@ public interface CommsNotification {
 
     // Diag Ping and pong message for displaying
     void diagMsg(MessageLedger msgLedger);
-
-
-    // error message from comms
-    void handleError(String errorMsg);
-
 }
